feat: US-006 - Create ScheduleSidebarItem component
Extract ScheduleSidebarItem into its own file at
frontend/src/components/sidebar/ScheduleSidebarItem.tsx
with status indicators, truncated names, and thread links.
