+++ frontend/src/components/drawers/app-sidebar.tsx
@@ -65,8 +65,7 @@ import useLinkClick from "@/hooks/useLinkClick";
 import { AxiosResponse } from "axios";
 import { useScheduleExecutions } from "@/hooks/useScheduleExecutions";
 import { useSchedules } from "@/hooks/useSchedules";
-import { ScheduleExecution } from "@/lib/entities/schedule";
-import { getExecutionStatusColor } from "@/lib/utils/calendar";
+import { ScheduleSidebarItem } from "@/components/sidebar/ScheduleSidebarItem";
 
 interface AssistantItemProps {
 	agent: Agent;
@@ -714,68 +713,6 @@ function CollapsibleGroup({
 				</CollapsibleContent>
 			</SidebarGroup>
 		</Collapsible>
-	);
-}
-
-interface ScheduleSidebarItemProps {
-	execution: ScheduleExecution;
-	scheduleName: string;
-}
-
-function ScheduleSidebarItem({
-	execution,
-	scheduleName,
-}: ScheduleSidebarItemProps) {
-	const statusEmoji =
-		execution.status === "success"
-			? "🟢"
-			: execution.status === "failure"
-				? "🔴"
-				: "🔵";
-	const statusColor = getExecutionStatusColor(execution.status);
-	const relativeTime = formatDistanceToNow(
-		new Date(execution.scheduled_time),
-		{ addSuffix: true },
-	);
-	const hasThread = !!execution.thread_id;
-
-	const handleClick = () => {
-		if (hasThread) {
-			window.open(`/thread/${execution.thread_id}`, "_blank", "noopener,noreferrer");
-		}
-	};
-
-	return (
-		<SidebarMenuItem className="mb-1">
-			<SidebarMenuButton
-				asChild
-				className={`h-auto px-3 py-2 rounded-lg border transition-all bg-transparent border-sidebar-border hover:bg-sidebar-accent/50 hover:border-sidebar-accent/50 ${
-					!hasThread ? "opacity-60 cursor-default" : "cursor-pointer"
-				}`}
-			>
-				<button
-					onClick={handleClick}
-					title={hasThread ? `Open thread` : "Execution pending"}
-				>
-					<div className="flex items-center gap-2 w-full min-w-0">
-						<span
-							className="shrink-0 text-xs"
-							style={{ color: statusColor }}
-						>
-							{statusEmoji}
-						</span>
-						<div className="flex flex-col min-w-0 flex-1">
-							<span className="text-sm font-medium text-sidebar-foreground truncate">
-								{scheduleName}
-							</span>
-							<span className="text-[10px] text-sidebar-foreground/50">
-								{relativeTime}
-							</span>
-						</div>
-					</div>
-				</button>
-			</SidebarMenuButton>
-		</SidebarMenuItem>
 	);
 }
